add text_proj_first option, fix bf16 timestep precision, and improve eval robustness
- add text_proj_first config for flux2_edit/mmdit_edit to control proj/norm order
- fix bf16 precision collapse in timestep computation (cfm.py, cfm_edit.py)
- add NaN/Inf detection and skip-existing in eval_infer_batch_edit.py
- add clean_ckpts.py script for checkpoint cleanup
- update CLAUDE.md with workflow guidelines

+++ src/f5_tts/model/cfm.py
@@ -208,10 +208,12 @@ class CFM(nn.Module):
             y0 = (1 - t_start) * y0 + t_start * test_cond
             steps = int(steps * (1 - t_start))
 
+        # Always compute timesteps in fp32 to avoid bf16 precision collapse
+        # (e.g. cos(small_x) rounds to 1.0 in bf16, making adjacent steps equal)
         if t_start == 0 and use_epss:  # use Empirically Pruned Step Sampling for low NFE
-            t = get_epss_timesteps(steps, device=self.device, dtype=step_cond.dtype)
+            t = get_epss_timesteps(steps, device=self.device, dtype=torch.float32)
         else:
-            t = torch.linspace(t_start, 1, steps + 1, device=self.device, dtype=step_cond.dtype)
+            t = torch.linspace(t_start, 1, steps + 1, device=self.device, dtype=torch.float32)
         if sway_sampling_coef is not None:
             t = t + sway_sampling_coef * (torch.cos(torch.pi / 2 * t) - 1 + t)
